refactor(parser): extract stage related rules.
because they are used both in query.rs and statement.rs
