+++ src/query/ast/src/parser/mod.rs
@@ -16,6 +16,7 @@ pub mod expr;
 #[allow(clippy::module_inception)]
 mod parser;
 pub mod query;
+mod stage;
 pub mod statement;
 pub mod token;
 pub mod unescape;
